Adding machineID,options columns to instance lists
